pyproject.toml
[project]
name = "nonebot-plugin-xibao"
version = "1.1.0"
description = "生成一张喜报（或者是悲报）！"
readme = "README.md"
requires-python = ">=3.9, <4.0"
authors = [{ name = "sp2dev", email = "theinsp2@outlook.com" }]
dependencies = [
  "nonebot2>=2.4.2,<3.0.0",
  "nonebot-plugin-send-anything-anywhere>=0.7.0",
  "skia-python>=130.0"
]


[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
